% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterio_informacion.R
\name{ValAIC}
\alias{ValAIC}
\title{Calcula el Criterio de Información de Akaike (AIC)}
\usage{
ValAIC(x, y, k)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}

\item{k}{Número de parámetros del modelo. Se utiliza para ajustar el AIC.}
}
\value{
Una lista con dos elementos:
\describe{
\item{AIC}{Vector con el coeficiente de información de Akaike para cada modelo.}
\item{AICc}{Vector con el coeficiente de información de Akaike corregido para cada modelo.}
}
}
\description{
Esta función calcula el coeficiente de información de Akaike (AIC) entre las predicciones y las observaciones.
El AIC es una medida que equilibra el ajuste del modelo con su complejidad y se define como:
\deqn{AIC = -2 \times \ell L + 2k,}
donde \eqn{\ell L} es la log-verosimilitud del modelo y \eqn{k} es el número de parámetros.
Además, se calcula el AIC corregido (AICc) para muestras pequeñas:
\deqn{AICc = AIC + \frac{2k(k+1)}{n - k - 1},}
donde \eqn{n} es el tamaño de la muestra.
}
\details{
La función une las observaciones y las predicciones en un solo objeto (omitiendo valores faltantes)
y, para cada modelo (columna de \code{x}), calcula el valor del AIC y el AIC corregido (AICc). Si el valor
de \code{k} no es compatible con el tamaño de la muestra, se detiene la ejecución.
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs <- data.frame(real1 = c(10, 11, 13))
  # Supongamos que el modelo tiene 3 parámetros:
  resultado <- ValAIC(pred, obs, k = 3)
  print(resultado$AIC)
  print(resultado$AICc)
}
}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
