% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencias.R
\name{ValANOVA}
\alias{ValANOVA}
\title{Realiza la Inferencia Paramétrica con ANOVA}
\usage{
ValANOVA(x, y)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}
}
\value{
Una lista en la que cada elemento es el resumen de la prueba ANOVA para cada modelo.
}
\description{
Esta función realiza una prueba ANOVA de una vía para comparar las predicciones y las observaciones.
Para cada modelo (columna de \code{x} y \code{y}), combina la información en un único data frame, añadiendo
un factor que indica si el valor proviene de las predicciones ("Pred") o de las observaciones ("Obs"), y luego
realiza la prueba ANOVA para determinar si existen diferencias significativas.
}
\details{
Para cada modelo, la función:
\enumerate{
\item Combina los valores de \code{x} y \code{y} en un único data frame.
\item Crea un factor que clasifica los valores en "Pred" y "Obs".
\item Realiza una prueba ANOVA de una vía para evaluar si existen diferencias significativas
entre las predicciones y las observaciones.
}
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs  <- data.frame(real1 = c(10, 11, 13), real2 = c(8, 10, 12))
  resultados <- ValANOVA(pred, obs)
  # Para ver el resumen de la prueba ANOVA del Modelo 1:
  print(resultados[["Modelo 1"]])
}

}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
