% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesgo_robustez.R
\name{ValMBD}
\alias{ValMBD}
\title{Calcula la Desviación Media del Sesgo (MBD)}
\usage{
ValMBD(x, y)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}
}
\value{
Una lista con un elemento:
\describe{
\item{mbd}{Matriz con la desviación media del sesgo para cada combinación de modelo y observación.}
}
}
\description{
Esta función calcula la desviación media del sesgo entre las predicciones y las observaciones.
El sesgo se mide como la diferencia promedio entre los valores predichos y los observados.
}
\details{
Para cada par de columnas correspondientes a un modelo, la función calcula
la diferencia \eqn{(x - y)} en cada posición y luego obtiene la media de estas diferencias,
proporcionando una medida del sesgo entre las predicciones y las observaciones.
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs  <- data.frame(real1 = c(8, 11, 13), real2 = c(10, 10, 16))
  resultado <- ValMBD(pred, obs)
  print(resultado$mbd)
}
}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})\cr
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}\cr
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
