% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencias.R
\name{ValMLE}
\alias{ValMLE}
\title{Calcula las Estimaciones de Máxima Verosimilitud (MLE)}
\usage{
ValMLE(x, y)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}
}
\value{
Una lista con dos elementos:
\describe{
\item{mu}{Vector con la media de la diferencia (predicciones - observaciones) para cada modelo.}
\item{sigma2}{Vector con la varianza de la diferencia para cada modelo.}
}
}
\description{
Esta función calcula las estimaciones de máxima verosimilitud para la diferencia entre las predicciones y las observaciones,
asumiendo que dicha diferencia sigue una distribución normal. Para cada modelo (columna en \code{x} y \code{y}),
se estima la media (\eqn{\mu}) y la varianza (\eqn{\sigma^2}) de la diferencia.
}
\details{
Para cada modelo, la función calcula la diferencia entre las predicciones y las observaciones,
y estima la media (\eqn{\mu}) y la varianza (\eqn{\sigma^2}) de dicha diferencia utilizando la función
\code{mean(..., na.rm = TRUE)}. Estas estimaciones se basan en la suposición de que la diferencia sigue
una distribución normal.
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs <- data.frame(real1 = c(8, 11, 13), real2 = c(10, 10, 15))
  resultado <- ValMLE(pred, obs)
  print(resultado$mu)
  print(resultado$sigma2)
}
}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
