% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_metricas_compuestas.R
\name{ValNash}
\alias{ValNash}
\title{Calcula el Índice de Eficiencia de Nash-Sutcliffe (Nash)}
\usage{
ValNash(x, y)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}
}
\value{
Una lista con un elemento:
\describe{
\item{Eficiencia}{Matriz con el índice de eficiencia de Nash-Sutcliffe para cada combinación de modelo y observación.}
}
}
\description{
Esta función calcula el índice de eficiencia de Nash-Sutcliffe entre las predicciones y las observaciones.
El índice de Nash-Sutcliffe (NSE) es una medida que evalúa la capacidad predictiva de un modelo; su valor varía entre
\eqn{-\infty} y 1, donde 1 indica una perfecta concordancia entre las predicciones y las observaciones.
}
\details{
Para cada par de columnas correspondientes a un modelo, la función calcula el índice de eficiencia de Nash-Sutcliffe
utilizando la fórmula:
\deqn{NSE = 1 - \frac{\sum_{i=1}^{n} (x_i - y_i)^2}{\sum_{i=1}^{n} (y_i - \bar{y})^2},}
donde \eqn{\bar{y}} es la media de las observaciones y \eqn{n} es el número de observaciones.
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs <- data.frame(real1 = c(10, 11, 13), real2 = c(8, 10, 12))
  resultado <- ValNash(pred, obs)
  print(resultado$Eficiencia)
}
}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
