% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errores_cuadraticos.R
\name{ValRRMSE}
\alias{ValRRMSE}
\title{Calcula el Error Cuadrático Medio Relativo (RRMSE)}
\usage{
ValRRMSE(x, y)
}
\arguments{
\item{x}{Data frame, matriz o vector numérico que contiene las predicciones.}

\item{y}{Data frame, matriz o vector numérico que contiene las observaciones.}
}
\value{
Una lista con un elemento:
\describe{
\item{rrmse}{Matriz con los errores cuadráticos medios relativos.}
}
}
\description{
Esta función calcula el error cuadrático medio relativo entre las predicciones y las observaciones,
dividiendo la raíz del error cuadrático medio (RMSE) por la media de las observaciones. Esta métrica permite evaluar la
magnitud del error en relación con los valores reales.
}
\details{
Para cada modelo (columna de \code{x}) y cada conjunto de observaciones (columna de \code{y}),
se calcula:
\deqn{RRMSE = \frac{\sqrt{\frac{1}{n}\sum_{i=1}^{n}(x_i - y_i)^2}}{\bar{y}}}
}
\examples{
\dontrun{
  # Ejemplo de uso:
  pred <- data.frame(modelo1 = c(10, 12, 14), modelo2 = c(9, 11, 15))
  obs <- data.frame(real1 = c(10, 11, 13))
  resultado <- ValRRMSE(pred, obs)
  print(resultado$rrmse)
}

#' @export

}
\author{
Henry P. Zumaeta Lozano (\email{henry.zumaeta.l@uni.pe})
LinkedIn: \href{https://www.linkedin.com/in/henryzumaeta}{henryzumaeta}
WhatsApp: \href{https://wa.me/51963719768}{+51963719768}
}
